requirements/faust.txt
# Stream Processing
# ------------------------------------------------------------------------------
faust-streaming==0.2.1
faust-streaming[uvloop]
pytz==2020.1
aiokafka
uvicorn==0.11.8  # https://github.com/encode/uvicorn
gevent==20.6.2
python-dateutil>=2.1.0